requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
pdfplumber==0.11.0
pymupdf==1.23.26
tiktoken==0.6.0
slowapi==0.1.9
structlog==24.1.0
httpx==0.27.0
numpy==1.26.4
